# GEODE-6218: Improves UTF-8 hashing function performance.