… cppcache/src/util/functional.cpp