* Update benchmark to measure different Unicode character widths.
* Reduces runtime by ~93%