@pivotal-jbarrett 2018-12-18
Original
```
2018-12-17 21:50:45
Running build/release/cppcache/benchmark/cpp-benchmark
Run on (12 X 2900 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 262K (x6)
  L3 Unified 12582K (x1)
-------------------------------------------------------------------------------------------------------
Benchmark                                                                Time           CPU Iterations
-------------------------------------------------------------------------------------------------------
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/8                   284 ns        283 ns    2521441
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/64                  541 ns        536 ns    1397206
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/512                2227 ns       2039 ns     351471
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/4096              15861 ns      14059 ns      49659
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/8192              36036 ns      27733 ns      25247
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/8                 11 ns          9 ns   79690346
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/64                65 ns         59 ns   11822128
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/512              482 ns        482 ns    1445063
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/4096            3949 ns       3917 ns     179744
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/8192            8065 ns       8057 ns      87723
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/8                387 ns        387 ns    2110557
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/64               726 ns        725 ns     984349
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/512             2980 ns       2962 ns     255702
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/4096           19848 ns      19770 ns      34141
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/8192           39398 ns      39237 ns      16477
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/8               7 ns          7 ns   97678053
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/64             58 ns         58 ns   12392451
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/512           500 ns        499 ns    1377573
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/4096         3994 ns       3993 ns     174216
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/8192         7973 ns       7950 ns      87496
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/8              318 ns        317 ns    2036186
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/64             659 ns        658 ns    1196995
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/512           2799 ns       2793 ns     242134
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/4096         19554 ns      19545 ns      35937
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/8192         38850 ns      38568 ns      17468
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/8             5 ns          5 ns  116656945
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/64           56 ns         56 ns   13241275
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/512         499 ns        498 ns    1431282
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/4096       4042 ns       4039 ns     177790
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/8192       8095 ns       8086 ns      86458
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/8                 503 ns        502 ns    1383946
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/64               1212 ns       1209 ns     542842
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/512              8325 ns       7633 ns      93064
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/4096            60274 ns      59837 ns      11612
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/8192           102605 ns     102223 ns       5838
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/8               16 ns         14 ns   62160890
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/64             121 ns        121 ns    5700419
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/512           1009 ns       1009 ns     691966
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/4096          8078 ns       8071 ns      86594
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/8192         16130 ns      15951 ns      43137
```

New:
```
2018-12-17 21:40:40
Running build/release/cppcache/benchmark/cpp-benchmark
Run on (12 X 2900 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 262K (x6)
  L3 Unified 12582K (x1)
-------------------------------------------------------------------------------------------------------
Benchmark                                                                Time           CPU Iterations
-------------------------------------------------------------------------------------------------------
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/8                    1 ns          9 ns  103649959
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/64                   49 ns         49 ns   13887787
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/512                 445 ns        445 ns    1585303
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/4096               3583 ns       3579 ns     204460
BM_Sequential<std::string, LATIN_CAPITAL_LETTER_C>/8192               6980 ns       6976 ns      98035
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/8                  5 ns          5 ns  142262254
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/64                50 ns         50 ns   13954230
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/512              450 ns        449 ns    1565894
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/4096            3541 ns       3540 ns     195406
BM_Sequential<std::u16string, LATIN_CAPITAL_LETTER_C>/8192            7104 ns       7100 ns      99364
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/8                 16 ns         16 ns   45245652
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/64               172 ns        172 ns    5700279
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/512             1068 ns       1068 ns     504515
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/4096            9065 ns       9063 ns      71803
BM_Sequential<std::string, INVERTED_EXCLAMATION_MARK>/8192           21053 ns      21044 ns      40003
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/8               9 ns          9 ns   79203440
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/64             65 ns         65 ns   11127007
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/512           623 ns        543 ns    1308558
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/4096         4379 ns       4308 ns     162651
BM_Sequential<std::u16string, INVERTED_EXCLAMATION_MARK>/8192         9689 ns       8651 ns      80488
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/8               34 ns         34 ns   20731097
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/64             266 ns        266 ns    2639219
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/512           2142 ns       2140 ns     321493
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/4096         16650 ns      16553 ns      43927
BM_Sequential<std::string, SAMARITAN_PUNCTUATION_ZIQAA>/8192         30605 ns      30279 ns      21789
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/8             8 ns          8 ns  103278350
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/64           59 ns         59 ns   11817537
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/512         524 ns        521 ns    1315666
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/4096       4262 ns       4220 ns     165392
BM_Sequential<std::u16string, SAMARITAN_PUNCTUATION_ZIQAA>/8192       8436 ns       8389 ns      83291
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/8                  45 ns         45 ns   15571577
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/64                298 ns        298 ns    2054805
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/512              2196 ns       2196 ns     280972
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/4096            18718 ns      18705 ns      45305
BM_Sequential<std::string, LINEAR_B_SYLLABLE_B008_A>/8192            35890 ns      35860 ns      18444
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/8               11 ns         11 ns   55602774
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/64             120 ns        120 ns    5857790
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/512           1033 ns       1030 ns     702120
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/4096          8273 ns       8244 ns      87712
BM_Sequential<std::u16string, LINEAR_B_SYLLABLE_B008_A>/8192         16668 ns      16625 ns      41981
```